Updated function name change and interface.

+++ arch/x86/lcd-domains/lcd-domains-arch.c
@@ -1176,7 +1176,7 @@ static void vmx_free_ept_dir_level(lcd_arch_epte_t *dir, int level)
 		for (idx = 0; idx < LCD_ARCH_PTRS_PER_EPTE; idx++) {
 			if (vmx_epte_present(dir[idx]))
 				printk(KERN_ERR "vmx_free_ept_dir_level: potential memory leak at hva %lx (hpa %lx)\n",
-					vmx_epte_hva(dir[idx]),
+					(unsigned long)vmx_epte_hva(dir[idx]),
 					__pa(vmx_epte_hva(dir[idx])));
 		}
 	} else {
